Fix scanf semantic change since V7
In V7, scanf would skip over the elements that didn't match. This
means that the scanf of %d%s%s%d wouldn't match most of the lines in
the config files because today's scanf won't skip elements, but
terminate the scanning.  Fix this by checking to see if the line
starts with a digit or not. If so, use one format, if not use another
and fake the missing data.